fix the sequence of tests
