+++ tests/00-solana-program.ts
@@ -1,0 +1,25 @@
+import * as anchor from "@coral-xyz/anchor";
+import { Program } from "@coral-xyz/anchor";
+import { SolanaProgram } from "../target/types/solana_program";
+
+import { expect } from "chai";
+
+describe("deploy-program", () => {
+  const provider = anchor.AnchorProvider.env();
+
+  anchor.setProvider(provider);
+
+  const program =
+    anchor.workspace.SolanaProgram as Program<SolanaProgram>;
+
+  it("Program is deployed on-chain", async () => {
+    const accountInfo =
+      await provider.connection.getAccountInfo(
+        program.programId
+      );
+
+    expect(accountInfo).to.not.be.null;
+
+    expect(accountInfo?.executable).to.be.true;
+  });
+});
